method_should_succeed moved to utility/utils
